Add support for stats command arg
This is an odd one.  The protocol documents that this is allowed, but then
goes on to describe that what to do with it is undefined:

> stats <args>\r\n
>
> Depending on <args>, various internal data is sent by the server. The
> kinds of arguments and the data sent are not documented in this version
> of the protocol, and are subject to change for the convenience of
> memcache developers.

In my tests so far memcached just returns an error.

So this adds support to the protocol layer, but at the server layer it
will only return an error.